tests: additional refactoring of login
