+++ features/home/100_anonymous_view.feature
@@ -3,15 +3,16 @@ Feature: An anonymous user can view the home page
   As an anonymous user
   I need to see a homepage
 
+  @smoke @anonymous
   Scenario: An anonymous user lands on the front page
     Given I am on the homepage
     Then I should see "Competency Frameworks"
     And I should see "About OpenSALT"
-    And I should see "Sign in" in the "a.login" element
+    And I should see "Sign in" in the header
 
     When I follow "About OpenSALT"
     Then I should see "Open Source"
-    And I should see "Sign in" in the "a.login" element
+    And I should see "Sign in" in the header
 
     When I follow "Sign in"
     Then I should see "Username"
