+++ tests/_support/AcceptanceTester.php
@@ -22,48 +22,63 @@ class AcceptanceTester extends \Codeception\Actor implements Context
 {
     use _generated\AcceptanceTesterActions;
 
-     /**
-      * Define custom actions here
-      */
-
-     /**
-      * @Given I am on the homepage
-      */
-     public function iAmOnTheHomepage()
-     {
+    /**
+     * @Given I am on the homepage
+     */
+    public function iAmOnTheHomepage(): AcceptanceTester
+    {
         $this->amOnPage('/');
-     }
 
-     /**
-      * @Then I should see :arg1
-      */
-     public function iShouldSee($arg1)
-     {
+        return $this;
+    }
+
+    /**
+     * @Then I should see :arg1
+     */
+    public function iShouldSee(string $arg1): AcceptanceTester
+    {
         $this->see($arg1);
-     }
 
-     /**
-      * @Then I should see :arg1 in the :arg2 element
-      */
-     public function iShouldSeeInTheElement($arg1, $arg2)
-     {
+        return $this;
+    }
+
+    /**
+     * @Then I should see :arg1 in the header
+     */
+    public function iShouldSeeInTheHeader(string $arg1): AcceptanceTester
+    {
+        $this->see($arg1, 'header');
+
+        return $this;
+    }
+
+    /**
+     * @Then I should see :arg1 in the :arg2 element
+     */
+    public function iShouldSeeInTheElement(string $arg1, string $arg2): AcceptanceTester
+    {
         $this->see($arg1, $arg2);
-     }
 
-     /**
-      * @When I follow :arg1
-      */
-     public function iFollow($arg1)
-     {
+        return $this;
+    }
+
+    /**
+     * @When I follow :arg1
+     */
+    public function iFollow(string $arg1): AcceptanceTester
+    {
         $this->click($arg1);
-     }
 
-     /**
-      * @When I press :arg1
-      */
-     public function iPress($link)
-     {
+        return $this;
+    }
+
+    /**
+     * @When I press :arg1
+     */
+    public function iPress(string $link): AcceptanceTester
+    {
         $this->click($link);
-     }
 
+        return $this;
+    }
 }
