+++ tests/_support/Page/Login.php
@@ -2,21 +2,25 @@
 
 namespace Page;
 
-class Login
+use Behat\Behat\Context\Context;
+use Doctrine\ORM\EntityManager;
+use Salt\UserBundle\Entity\User;
+use Salt\UserBundle\Entity\Organization;
+use Salt\UserBundle\Repository\UserRepository;
+
+class Login implements Context
 {
     // include url of current page
     public static $loginUrl = '/login';
     public static $logoutUrl = '/logout';
 
-    /**
-     * Declare UI map for this page here. CSS or XPath allowed.
-     * public static $usernameField = '#username';
-     * public static $formSubmitButton = "#mainForm input[type=submit]";
-     */
     public static $loginLink = 'a.login';
     public static $usernameField = '#username';
     public static $passwordField = '#password';
     public static $loginButton = 'button.btn-primary[type=submit]';
+
+    protected static $users = [];
+    protected static $lastUser = null;
 
     /**
      * @var \AcceptanceTester
@@ -28,6 +32,10 @@ class Login
         $this->I = $I;
     }
 
+    /**
+     * @Given I logout
+     * @Given I am logged out
+     */
     public function logout(): Login
     {
         $I = $this->I;
@@ -37,10 +45,16 @@ class Login
         return $this;
     }
 
+    /**
+     * @Given I log in as a user with role :role
+     * @Given I log in as a :role user
+     */
     public function loginAsRole(string $role): Login
     {
-        $user = $this->getUserWithRole($role);
-        $this->loginWithPassword($user['username'], $user['password']);
+        $this
+            ->logout()
+            ->aUserExistsWithRole($role)
+            ->loginWithPassword(self::$lastUser['username'], self::$lastUser['password']);
 
         return $this;
     }
@@ -58,6 +72,87 @@ class Login
         $I->seeLink('Logout');
 
         $I->iAmOnTheHomepage();
+
+        return $this;
+    }
+
+    /**
+     * @When I fill in the username
+     */
+    public function iFillInTheUsername(): Login
+    {
+        $this->I->fillField(self::$usernameField, self::$lastUser['user']);
+
+        return $this;
+    }
+
+    /**
+     * @When I fill in the password
+     */
+    public function iFillInThePassword(): Login
+    {
+        $this->I->fillField(self::$passwordField, self::$lastUser['pass']);
+
+        return $this;
+    }
+
+    /**
+     * @Given a user exists with role :role
+     */
+    public function aUserExistsWithRole(string $role): Login
+    {
+        /** @var EntityManager $em */
+        $em = $this->I->grabService('doctrine.orm.default_entity_manager');
+
+        /** @var \Faker\Generator $faker */
+        $faker = \Faker\Factory::create();
+
+        $role = preg_replace('/[^A-Z]/', '_', strtoupper($role));
+        $password = $faker->password;
+
+        /** @var UserRepository $userRepo */
+        $userRepo = $em->getRepository(User::class);
+        $user = $userRepo->createQueryBuilder('u')
+            ->where('u.username like :prefix')
+            ->setParameter(':prefix', 'TEST:'.$role.':%')
+            ->setMaxResults(1)
+            ->getQuery()
+            ->getOneOrNullResult();
+
+        if ($user) {
+            $username = $user->getUsername();
+            $userRepo->setUserPassword($username, $password);
+        } else {
+            $orgRepo = $em->getRepository(Organization::class);
+            $org = $orgRepo->createQueryBuilder('o')
+                ->where('o.name like :prefix')
+                ->setParameter(':prefix', 'TEST:%')
+                ->setMaxResults(1)
+                ->getQuery()
+                ->getOneOrNullResult();
+            if (!$org) {
+                $org = $orgRepo->addNewOrganization(
+                    'TEST:'.$faker->company
+                );
+            }
+
+            $username = 'TEST:'.$role.':'.$faker->userName;
+            $userRepo->addNewUser($username, $org, $password, $role);
+
+            $user = $userRepo->createQueryBuilder('u')
+                ->where('u.username = :username')
+                ->setParameter(':username', $username)
+                ->setMaxResults(1)
+                ->getQuery()
+                ->getOneOrNullResult();
+
+            $this->I->assertNotEmpty($user, 'User could not be created.');
+        }
+
+        $this->I->comment('I find the username and password');
+
+        self::$lastUser = ['user' => $username, 'pass' => $password];
+        self::$users[] = self::$lastUser;
 
         return $this;
     }
