+++ tests/acceptance.suite.yml
@@ -15,6 +15,7 @@ modules:
             host: 'phantomjs'
             port: 8643
         - \Helper\Acceptance
+        - Asserts
 
 paths:
     envs: tests/_envs
@@ -23,4 +24,4 @@ gherkin:
     contexts:
         default:
             - AcceptanceTester
-            - "Context\\Login"
+            - "Page\\Login"
